+++ aquadoggo/src/db/stores/operation.rs
@@ -98,11 +98,11 @@ impl OperationStore<VerifiedOperation> for SqlStorage {
         .bind(document_id.as_str())
         .bind(operation.operation_id().as_str())
         .bind(operation.action().as_str())
-        .bind(operation.schema().as_str())
+        .bind(operation.schema().to_string())
         .bind(
             operation
                 .previous_operations()
-                .map(|document_view_id| document_view_id.as_str()),
+                .map(|document_view_id| document_view_id.to_string()),
         )
         .execute(&self.pool)
         .await
@@ -294,6 +294,7 @@ mod tests {
     };
     use rstest::rstest;
 
+    use crate::db::provider::SqlStorage;
     use crate::db::stores::test_utils::{test_db, TestDatabase, TestDatabaseRunner};
 
     #[rstest]
@@ -309,7 +310,7 @@ mod tests {
         document_id: DocumentId,
         #[from(test_db)] runner: TestDatabaseRunner,
     ) {
-        runner.with_db_teardown(|db: TestDatabase| async move {
+        runner.with_db_teardown(|db: TestDatabase<SqlStorage>| async move {
             // Construct the storage operation.
             let operation = VerifiedOperation::new(&author, &operation_id, &operation).unwrap();
 
@@ -337,7 +338,7 @@ mod tests {
         document_id: DocumentId,
         #[from(test_db)] runner: TestDatabaseRunner,
     ) {
-        runner.with_db_teardown(|db: TestDatabase| async move {
+        runner.with_db_teardown(|db: TestDatabase<SqlStorage>| async move {
             db.store
                 .insert_operation(&verified_operation, &document_id)
                 .await
@@ -362,7 +363,7 @@ mod tests {
         document_id: DocumentId,
         #[from(test_db)] runner: TestDatabaseRunner,
     ) {
-        runner.with_db_teardown(|db: TestDatabase| async move {
+        runner.with_db_teardown(|db: TestDatabase<SqlStorage>| async move {
             assert!(db
                 .store
                 .get_document_by_operation_id(create_operation.operation_id())
@@ -407,7 +408,7 @@ mod tests {
         #[with(5, 1, 1)]
         runner: TestDatabaseRunner,
     ) {
-        runner.with_db_teardown(|db: TestDatabase| async move {
+        runner.with_db_teardown(|db: TestDatabase<SqlStorage>| async move {
             let author = Author::try_from(key_pair.public_key().to_owned()).unwrap();
 
             let latest_entry = db