+++ aquadoggo/src/errors.rs
@@ -1,4 +1,4 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
 
-/// A specialized result type for the storage provider.
-pub type StorageProviderResult<T> = anyhow::Result<T, Box<dyn std::error::Error + Send + Sync>>;
+/// A result type used in aquadoggo modules.
+pub type Result<T> = anyhow::Result<T, Box<dyn std::error::Error + Send + Sync>>;