+++ aquadoggo/src/materializer/tasks/dependency.rs
@@ -187,6 +187,7 @@ mod tests {
 
     use crate::config::Configuration;
     use crate::context::Context;
+    use crate::db::provider::SqlStorage;
     use crate::db::stores::test_utils::{
         insert_entry_operation_and_view, send_to_store, test_db, TestDatabase, TestDatabaseRunner,
     };
@@ -306,7 +307,7 @@ mod tests {
         #[case] runner: TestDatabaseRunner,
         #[case] expected_next_tasks: usize,
     ) {
-        runner.with_db_teardown(move |db: TestDatabase| async move {
+        runner.with_db_teardown(move |db: TestDatabase<SqlStorage>| async move {
             let context = Context::new(
                 db.store.clone(),
                 Configuration::default(),
@@ -346,7 +347,7 @@ mod tests {
         #[with(1, 1, 1)]
         runner: TestDatabaseRunner,
     ) {
-        runner.with_db_teardown(|db: TestDatabase| async move {
+        runner.with_db_teardown(|db: TestDatabase<SqlStorage>| async move {
             let context = Context::new(
                 db.store.clone(),
                 Configuration::default(),
@@ -410,7 +411,7 @@ mod tests {
         #[case] document_view_id: Option<DocumentViewId>,
         #[from(test_db)] runner: TestDatabaseRunner,
     ) {
-        runner.with_db_teardown(|db: TestDatabase| async move {
+        runner.with_db_teardown(|db: TestDatabase<SqlStorage>| async move {
             let context = Context::new(
                 db.store,
                 Configuration::default(),
@@ -456,7 +457,7 @@ mod tests {
         )
     )]
     fn fails_on_deleted_documents(#[case] runner: TestDatabaseRunner) {
-        runner.with_db_teardown(|db: TestDatabase| async move {
+        runner.with_db_teardown(|db: TestDatabase<SqlStorage>| async move {
             let context = Context::new(
                 db.store.clone(),
                 Configuration::default(),
@@ -493,7 +494,7 @@ mod tests {
         ])]
         runner: TestDatabaseRunner,
     ) {
-        runner.with_db_teardown(|db: TestDatabase| async move {
+        runner.with_db_teardown(|db: TestDatabase<SqlStorage>| async move {
             let context = Context::new(
                 db.store.clone(),
                 Configuration::default(),
@@ -578,7 +579,7 @@ mod tests {
         ])]
         runner: TestDatabaseRunner,
     ) {
-        runner.with_db_teardown(move |db: TestDatabase| async move {
+        runner.with_db_teardown(move |db: TestDatabase<SqlStorage>| async move {
             let context = Context::new(
                 db.store.clone(),
                 Configuration::default(),