+++ aquadoggo/src/replication/service.rs
@@ -5,21 +5,17 @@ use std::time::Duration;
 
 use anyhow::{anyhow, Result};
 use bamboo_rs_core_ed25519_yasmf::verify::verify_batch;
-use futures::TryFutureExt;
 use log::{debug, error, trace, warn};
 use p2panda_rs::entry::LogId;
 use p2panda_rs::entry::SeqNum;
 use p2panda_rs::identity::Author;
-use p2panda_rs::operation::{AsVerifiedOperation, VerifiedOperation};
-use p2panda_rs::storage_provider::traits::{
-    AsStorageEntry, EntryStore, OperationStore, StorageProvider,
-};
+use p2panda_rs::storage_provider::traits::{AsStorageEntry, EntryStore};
 use tokio::task;
 
 use crate::bus::{ServiceMessage, ServiceSender};
 use crate::context::Context;
-use crate::db::request::PublishEntryRequest;
 use crate::db::stores::StorageEntry;
+use crate::domain::publish;
 use crate::graphql::replication::client;
 use crate::manager::{ServiceReadySender, Shutdown};
 
@@ -160,66 +156,29 @@ async fn insert_new_entries(
     tx: ServiceSender,
 ) -> Result<()> {
     for entry in new_entries {
-        // Parse and validate parameters
-        let args = PublishEntryRequest {
-            entry: entry.entry_signed().clone(),
-            // We know a storage entry has an operation so we safely unwrap here.
-            operation: entry.operation_encoded().unwrap().clone(),
-        };
-
         // This is the method used to publish entries arriving from clients. They all contain a
         // payload (operation).
         //
         // @TODO: This is not a great fit for replication, as it performs validation we either do
-        // not need or already done in a previous step. We plan to refactor this into a more
+        // not need or have already done in a previous step. We plan to refactor this into a more
         // modular set of methods which can definitely be used here more cleanly. For now, we do it
         // this way.
-        context
-            .0
-            .store
-            .publish_entry(&args)
-            .await
-            .map_err(|err| anyhow!(format!("Error inserting new entry into db: {:?}", err)))?;
+        //
+        // @TODO: Additionally, when we implement payload deletion and partial replication we will
+        // be expecting entries to arrive here possibly without payloads.
 
-        // @TODO: We have to publish the operation too, once again, this will be improved with the
-        // above mentioned refactor.
-        let document_id = context
-            .0
-            .store
-            .get_document_by_entry(&entry.hash())
-            .await
-            .map_err(|err| anyhow!(format!("Error retrieving document id from db: {:?}", err)))?;
+        publish(
+            &context.0.store,
+            entry.entry_signed(),
+            entry
+                .operation_encoded()
+                .expect("All stored entries contain an operation"),
+        )
+        .await
+        .map_err(|err| anyhow!(format!("Error inserting new entry into db: {:?}", err)))?;
 
-        match document_id {
-            Some(document_id) => {
-                let operation = VerifiedOperation::new_from_entry(
-                    entry.entry_signed(),
-                    entry.operation_encoded().unwrap(),
-                )
-                // Safely unwrap here as the entry and operation were already validated.
-                .unwrap();
-
-                context
-                    .0
-                    .store
-                    .insert_operation(&operation, &document_id)
-                    .map_ok({
-                        let entry = entry.clone();
-                        let tx = tx.clone();
-
-                        move |_| {
-                            send_new_entry_service_message(tx.clone(), &entry);
-                        }
-                    })
-                    .map_err(|err| {
-                        anyhow!(format!("Error inserting new operation into db: {:?}", err))
-                    })
-                    .await
-            }
-            None => Err(anyhow!(
-                "No document found for published operation".to_string()
-            )),
-        }?;
+        // Send new entry & operation to other services.
+        send_new_entry_service_message(tx.clone(), entry);
     }
 
     Ok(())
@@ -252,7 +211,7 @@ async fn add_certpool_to_entries_for_verification(
     Ok(())
 }
 
-/// Helper method to inform other services (like materialization service) about new operations.
+/// Helper method to inform other services (like materialisation service) about new operations.
 fn send_new_entry_service_message(tx: ServiceSender, entry: &StorageEntry) {
     let bus_message = ServiceMessage::NewOperation(entry.entry_signed().hash().into());